TestSonar adapt the level
main: in Sonar.cpp moltiplico per 1000 perchè altrimenti ritorna valori sotto lo 0 che in differenza con WLMAX a 1000 mantengono sempre lo stato di allarme

add print
set level in Parameter at 1000, 800, 600
use float and not double

SONAR OK...
only to remove comment

+++ iot-assignment-2-pio/lib/Sonar/src/Sonar.h
@@ -14,7 +14,7 @@ class Sonar {
  private:
   int trigPin;
   int echoPin;
-  const double vs = 331.45 + 0.62 * 20;
+  const float vs = 331.45 + 0.62 * 20;
 };
 
 #endif
